Fix napari/napari#5183
Note that this is far from the only broken link, in the same issue, the
screenshot shows "as a Python package instead" that should also be a
link.

A quick way to see the number of non-existing links woudl be:

    $ rg 'class="xref myst"' | cut -f1 -d: |cut -f1-2 -d'/' | uniq -c | sort -n
       1 0.4.16/plugins
       1 dev/plugins
       2 0.4.15/index.html
       2 0.4.16/index.html
       2 0.4.16/tutorials
       4 0.4.16/roadmaps
       5 0.4.16/plugins
       6 0.4.15/plugins
       7 0.4.15/tutorials
       8 0.4.16/roadmaps
       9 0.4.15/developers
       9 0.4.16/developers
      12 0.4.15/guides
      12 0.4.15/roadmaps
      12 0.4.16/guides

Therefore there are potentially plenty of other links to fix in already
released version.