[HUGE DIFF] Fix version warning for stable & all previous versions (#437)
I swear this is the last time 😭

Most of these involved forcing the version number to avoid the issues with the files that were unversioned and got copied over. In that case, a mismatch between the version number for which the page was built and the version number for the pages where it lived was causing the bad version warning.

It's many files, but most changes are one (maybe two) lines per file.